added manualscale test (#3231)
