+++ test/functional/corerp/resources/container_test.go
@@ -86,3 +86,71 @@ func Test_ContainerHttpRoute(t *testing.T) {
 
 	test.Test(t)
 }
+
+func Test_ContainerReadinessLiveness(t *testing.T) {
+	template := "testdata/corerp-resources-container-liveness-readiness.bicep"
+	name := "corerp-resources-container-live-ready"
+
+	requiredSecrets := map[string]map[string]string{}
+
+	test := corerp.NewCoreRPTest(t, name, []corerp.TestStep{
+		{
+			Executor: step.NewDeployExecutor(template, functional.GetMagpieImage()),
+			CoreRPResources: &validation.CoreRPResourceSet{
+				Resources: []validation.CoreRPResource{
+					{
+						Name: name,
+						Type: validation.ApplicationsResource,
+					},
+					{
+						Name: "ctnr-live-ready",
+						Type: validation.ContainersResource,
+					},
+				},
+			},
+			K8sObjects: &validation.K8sObjectSet{
+				Namespaces: map[string][]validation.K8sObject{
+					"default": {
+						validation.NewK8sPodForResource(name, "ctnr-live-ready"),
+					},
+				},
+			},
+		},
+	}, requiredSecrets)
+
+	test.Test(t)
+}
+
+func Test_ContainerManualScale(t *testing.T) {
+	template := "testdata/corerp-azure-container-manualscale.bicep"
+	name := "corerp-resources-container-manualscale"
+
+	requiredSecrets := map[string]map[string]string{}
+
+	test := corerp.NewCoreRPTest(t, name, []corerp.TestStep{
+		{
+			Executor: step.NewDeployExecutor(template, functional.GetMagpieImage()),
+			CoreRPResources: &validation.CoreRPResourceSet{
+				Resources: []validation.CoreRPResource{
+					{
+						Name: name,
+						Type: validation.ApplicationsResource,
+					},
+					{
+						Name: "ctnr-manualscale",
+						Type: validation.ContainersResource,
+					},
+				},
+			},
+			K8sObjects: &validation.K8sObjectSet{
+				Namespaces: map[string][]validation.K8sObject{
+					"default": {
+						validation.NewK8sPodForResource(name, "ctnr-manualscale"),
+					},
+				},
+			},
+		},
+	}, requiredSecrets)
+
+	test.Test(t)
+}
