--- test/functional/corerp/resources/readinessliveness_test.go
@@ -1,49 +1,0 @@
-// ------------------------------------------------------------
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-// ------------------------------------------------------------
-
-package resource_test
-
-import (
-	"testing"
-
-	"github.com/project-radius/radius/test/functional"
-	"github.com/project-radius/radius/test/functional/corerp"
-	"github.com/project-radius/radius/test/step"
-	"github.com/project-radius/radius/test/validation"
-)
-
-func Test_ContainerReadinessLiveness(t *testing.T) {
-	template := "testdata/corerp-resources-container-liveness-readiness.bicep"
-	name := "corerp-resources-container-live-ready"
-
-	requiredSecrets := map[string]map[string]string{}
-
-	test := corerp.NewCoreRPTest(t, name, []corerp.TestStep{
-		{
-			Executor: step.NewDeployExecutor(template, functional.GetMagpieImage()),
-			CoreRPResources: &validation.CoreRPResourceSet{
-				Resources: []validation.CoreRPResource{
-					{
-						Name: name,
-						Type: validation.ApplicationsResource,
-					},
-					{
-						Name: "ctnr-live-ready",
-						Type: validation.ContainersResource,
-					},
-				},
-			},
-			K8sObjects: &validation.K8sObjectSet{
-				Namespaces: map[string][]validation.K8sObject{
-					"default": {
-						validation.NewK8sPodForResource(name, "ctnr-live-ready"),
-					},
-				},
-			},
-		},
-	}, requiredSecrets)
-
-	test.Test(t)
-}
